backend/requirements.txt
africastalking==1.2.0
amqp==2.5.2
asgiref==3.2.3
autopep8==1.5
billiard==3.6.3.0
cachetools==4.1.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorclass==2.2.0
contextlib2==0.6.0.post1
cryptography==2.8
Django==3.0.3
django-cacheops==4.2
django-cors-headers==3.2.1
django-redis==4.11.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docopt==0.6.2
drf-writable-nested==0.5.4
factory-boy==2.12.0
Faker==4.0.3
funcy==1.14
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
kombu==4.6.8
packaging==20.1
pip-upgrader==1.4.15
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
rsa==4.0
schema==0.7.2
six==1.14.0
sqlparse==0.3.0
terminaltables==3.1.0
text-unidecode==1.3
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
